+++ app/src/components/common/list_item/index.tsx
@@ -167,7 +167,7 @@ export const ListItem: React.FC<Props> = (props: Props) => {
     }
 
     setToken()
-  }, [collateralToken, account])
+  }, [collateralToken, account, provider, collateralVolume])
 
   const percentages = calcPrice(outcomeTokenAmounts)
   const indexMax = percentages.indexOf(Math.max(...percentages))
