+++ app/src/components/market/steps/index.ts
@@ -1,5 +1,4 @@
 export { AskQuestionStep } from './items/ask_question_step'
 export { FundingAndFeeStep } from './items/funding_and_fee_step'
-export { CreateMarketStep } from './items/create_market_step'
 export { SummaryMarketStep } from './items/summary_market_step'
 export { MenuStep } from './menu_step'
